onnnxruntime, python3Packages.onnxruntime: improve packaging
The Python bindings to onnxruntime were added by me in #193188.

Adding Python support this way is not a good way. Here a wheel was
created (which is fine) and installed in a python output. The
propagated-build-inputs file is put in the dev output. That's fine,
except that stdenv.mkDerivation does not automatically add dev when
python is included, because other outputs are only added when no
output is explicitly selected. This means that when you want to use
these bindings in another Python package, pip will complain it cannot
find the dependencies of the bindings.

In this PR, the onnxruntime derivation outputs a wheel in the dist output.
Then, in python-packages.nix we have a separate onnxruntime package
which installs the bindings.

The Python bindings have quite some dependencies which, depending on
your use case, are not required. Thus the dependency relax hook is
used to remove some of these dependencies.

Note there is also an issue with protobuf versions. The onnxruntime
bindings require an older protobuf and Python protobuf which we
cannot offer. Thus protobuf is also removed as Python dependency.

(cherry picked from commit 34d1d336adec65ecd6547924c5149709eec798de)