anki-bin: fix SSL errors in certain addons
One of the Anki addons I've used [1] encounters SSL cert verification
errors when ran on NixOS. The issue seems to be an interaction between
the URL request library the addon uses and the state of the FHS
environment.

Exporting `SSL_CERT_FILE` to point to the /etc-wide SSL file seems to
fix the issue. This seems to me on the surface to be a very non-nix way
of approaching the issue, however, ca.nix [2] seemingly implies to me
that this usage is OK and expected by certain software.

[1] https://ankiweb.net/shared/info/858591644
[2] https://github.com/NixOS/nixpkgs/blob/master/nixos/modules/security/ca.nix