libnfs: allow building with pthread support
libnfs is not thread-safe by default. Making it thread-safe requires building
it with pthread support and enabling multi-threading at runtime using
nfs_mt_service_thread_start() function.

The patch allows to optionally build libnfs with multi-threading support.

This is not enabled by default to avoid runtime costs for programs that don't
require libnfs to be thread-safe, so a separate top level package 'libnfs-pthread'
is added.