fortune-mod: create fortuneExtensions
The fortune-mod program is a true unix classic, being 43 years old,
originally released in 1979.

It has a significant cultural relevance in the unix sphere.

As such it has in it's long history had several extensions made.

This pull requests seeks to create a framework in nixpkgs for dealing
with these extensions.

What has been done:
- created most code for handling extensions
- created the blag-fortune extension (not done)

The blag-fortune extension serves only to be a demo to be an example for
other contributers as to how one would package an extension to
fortune-mod

This is still a work in progress, and I would be grateful if anyone more
experienced with the nix language would help me get this fully working.