libraspberrypi: build all binaries on 32-bit ARM
This package included a patch which accidently disabled the applications which
can only be built for 32-bit ARM on all platforms (including 32-bit ARM). There
is a newer version of the patch available in the upstream PR which can avoid
this problem, but I don't think the patch is truly necessary at all. We can
simply pass -DARM64=ON on all platforms except 32-bit ARM to get the same
effect.