hy: fix passthru.withPackages
In the previous version it was adding the extra packages as
`propagatedBuildInputs`. This meant that this package would be rebuild
for no reason (it is not like the package will actually depend on the
extra inputs) and also cause the strange side-effect of creating a hy
package without its console entry-points (e.g.: no `$out/bin` contents).

Now we are reusing the `python.withPackages` that will avoid the
unnecessary rebuild, fixing both issues.