update-python-libraries: fix namespaced fetchers
Some package expressions look like this:

	python3Packages.buildPythonApplication rec {
	  pname = "patatt";
	  version = "0.6.2";

	  src = python3Packages.fetchPypi {

These expressions, where fetchPypi is not a top-level binding, would
end up with an updateScript provided by buildPythonApplication that
would always fail, because update-python-libraries couldn't handle the
"python3Packages." prefix to the fetcher.  And because the
updateScript failed, nixpkgs-update wouldn't fall back to its default
update logic, so r-ryantm would never update these packages.

Here I've fixed that with a regular expression.  It would be nicer if
we could just instantiate the expression and look at the URL rather
than trying to parse the expression, but I don't have the capacity to
look into a big change like that at the moment.