cbqn: always compile using clang
Investigating the build failures of CBQN on i686 and aarch64, turns up
that CBQN is broken on these platforms only with gcc. In the case of
i686 due to a codegen bug, on aarch64 gcc doesn't like some CPP macros
CBQN uses.

The deeper reason for this is that CBQN primarily is developed for and
tested with clang, so it probably saves us trouble just using clang, as
it is no real problem for us to do so (clangStdenv is quite
unproblematic for pure C builds). This should also yield faster
binaries, since CBQN apparently uses some clang-specific tricks to get
quicker code generation.