pantheon.gala: 6.3.1 -> 7.0.1
We are delaying gala updates since Oct 2022 (6.3.2) because
those versions are not shipped in elementary OS 6.1 and not
many users test them and I decided I cannot properly test them.
So instead, we manually backport patches that we can confirm
fixes issues people care. We mostly manage to do that.

Starting from 7.0.0, this is shipped in the new elementary OS 7
stable release and used by more people. It still contains some
regressions compared to our patched 6.3.1 but should be usuable
in most cases (when you only have a single display). And starting
from 7.0.1, support for older mutter is dropped, and some of
the fixes are done together with large sized formatting changes,
making manual backport much more harder. So I decided that this
will be a good chance to ship this to NixOS users.