nixosTests.nscd: drop _gateway lookups
This has shown to be flaky in the VM test, at least when running on
the aarch64 ofborg builder(s).

I assume it's some flakyness in systemd-networkd not being fully up, or
at least not up to the point that it properly replies to the _gateway
request.

This part of the test is supposed to test external (non-glibc) nss
module lookup for the host database works, which is already sufficiently
covered in the previous checks (for *.localhost). Drop these redundant
checks. We're not integration-testing networkd here.