tor-browser-bundle-bin: add mirror for old versions
The normal source download mirrors usually contain only the latest (or a
few latest) release. Building a derivation for an old version can
therefore sometimes fail, if the source tar file was removed from the
mirror in the meantime.

A mirror containing even the old TBB release files is added to fix this
issue.