deadd-notification-center: 2022-04-20 -> unstable-2022-11-07
Update deadd-notification-center to the latest revision. Change the
expression from using the prebuilt binary provided in the repository to
building from source instead as well as provide a functioning systemd
service file as the one provided by upstream is broken.

Remove pacman99 from and add Melkor333 to maintainers as requested.