stdenv: fix cc for pseudo-crosscompilation
It is possible to force nixpkgs to take cross-compilation pathways,
e.g. using pkgsCross.gnu64 on an x86_64-linux system.

This final stdenv stage is predicated on `crossSystem == localSystem`,
but in a pseudo-crosscompilation scenario, the binaries are still
targetPrefixed even though the cross and local systems are the same.
This results in a nonfunctional cc-wrapper, usually manifesting as
`configure: error: C compiler cannot create executables`

Thus, this condition should be based on `targetPrefix == ""`
to handle that edge case.