lisp-modules-new: strip unrelated defsystems
Strip any (asdf:defsystem .. that is not for the specific system we want to
build. Fixes at least cl-async, which has a .asd with multiple defsystems, which
causes problems as the wrong (non-precompiled) package is loaded at build time,
which then tries to compile after the fact, putting a .fasl in a read-only
location (the store), breaking the system.