ghidra: add support for aarch64-darwin
This also changes the method used to change Java's notion of the user
home directory. Instead of redirecting /etc/passwd to a generated file,
the JAVA_TOOL_OPTIONS environment variable is used to provide a value
for the user.home system property. The redirect method did not seem to
be effective on my Monterey (12.6) system and the new approach is IMO
simpler anyway.