Dissallow underscore in python pkg attribute
This is a proposal to remove all underscores from python packages
internal attrNames, so have a standardized naming convention and not
the current mismash of flask_mail and flask-mailman for instance.

This PR will need more work I'm sure, but I'm opening this now from the
Nixcon hackathon so we can discuss and improve. I've talked with @fridh
and @Ericson2314 about this change in person at the conference.