flake.nix: NixOS `nixpkgs` registry pin as `nixos`
On a given NixOS system, if a user discovers the `nix.registry` option,
they may wish to set the nixpkgs of their system as a pin for
convenience, and rightfully so.

However, pinning the global nixpkgs to a `/nix/store` path can wreak
havoc with the lockfile of any other flake with an indirect reference
to nixpkgs, borking it for anyone who tries to call it remotely.

This is simply bad UX. It should be hard to do the wrong thing, but
this is nearly inevitable.

Instead of expecting the user to discover this on their own, we should
offer them the convenience of a pinning of their system's nixpkgs as
`nixos` and encourage them to use it as their "goto" nixpkgs, analogous
to `<nixpkgs>` in "pre-flake" material, but more precise.

Since we can expect this reference to exist we can always point to it
in documentation and reference material.