nextcloud25: init
This is a non-breaking change because the default version installed by
`services.nextcloud` remains `pkgs.nextcloud24`, but it gives users a
chance to upgrade to v25 on 22.05.

I also left `pkgs.nextcloud23` in here because it will only be dropped
in 2022-12 which is also the EOL of 22.05.

The corresponding change in `master` is at
40b7f52b8f4fd63393c1fdb838f022a154a4f2b2 from PR #197386.