zfsUnstable: 2.1.6 → 2.1.7-staging
With 5.19 kernels now being EoL and my laptop demanding at least 5.17 to
function, I want to get ahead of the curve and add support for 6.0
kernels. The 2.1.6 tag was cut so late that it support 6.0 but isn’t as
well tested according to the conversations I read. There aren’t many
commits in the `2.1.7-staging` at this point, but this will help keep my
system afloat until the real `2.1.7` release. I’ve booted my laptop into
this and everything appears to be functioning normally.