ruby: refactor choice of patches to apply
Now instead of always fetching the `head` patches for a `majMin`
version, we fetch patches specific to a `majMinTiny` version. At the
current commit of `rvm-patchsets` and current ruby versions, these are
exactly the same patches.

This doesn't change any behavior, but the checksum for the ruby
artifacts changes. I think it's because the patch files timestamps are
different.

This allows `mkRuby` to build many more versions of ruby instead of just
the latest patch.