retroarch: drop all Darwin related code
Darwin support in retroarch was broken for a long time, and it is
getting worse with each subsequent update. Upstream abandoned
`./configure` usage in macOS systems too for a Xcode build system.

So if we want to get back decent support for retroarch in Darwin, we
should migrate it to use Xcode. For now, let's just remove all this code
that is not working.