librewolf: build with `MOZ_REQUIRE_SIGNING = ""`
this allows sideloading unsigned browser extensions (via the
`nixExtensions` option in the [firefox-wrapper]). i believe librewolf
is meant to allow unsigned sideloading, based on upstream discussions
and this bit that already exists in ./librewolf.nix:

```nix
extraConfigureFlags = [
    "--with-unsigned-addon-scopes=app,system"
    "--allow-addon-sideload"
];
```

but Mozilla changed some defaults over the years. this `MOZ_REQUIRE_SIGNING`
option is mentioned [upstream] and we can see that the arch [[pkgbuild]]
clears it in a similar manner.

[upstream]: https://gitlab.com/librewolf-community/browser/arch/-/issues/49
[pkgbuild]: https://gitlab.com/librewolf-community/browser/arch/-/blob/master/PKGBUILD#L95
[firefox-wrapper]: pkgs/applications/networking/browsers/firefox/wrapper.nix