adoptopenjdk 17: remove (again)
this was originally replaced with temurin in c74221839b5, which landed
in staging in acf46b05697 (#140364)

but it was also added in b6cb656c0d8 (in support of openjdk 18 in
da40a449e9e), which landed directly on master in 9413ebb2e1d
(#165354).

those two conflicted when master was merged into staging-next in
a5dfac88d5893514bcd204ed7f820a0f42ac177c (#191339), and adoptopenjdk
17 was mistakenly kept during the conflict resolution.

the net result is that one would get:

$ nix-build -A pkgs.adoptopenjdk-hotspot-bin-17
error: Alias adoptopenjdk-hotspot-bin-17 is still in all-packages.nix

... instead of the desired:

$ nix-build -A pkgs.adoptopenjdk-hotspot-bin-17
error: AdoptOpenJDK is now Temurin. Use temurin-bin-17