ldmud: fix undefined reference to 'crypt'
Adding `libxcrypt` as a dependency is required to avoid a link time
error from an undefined reference to `crypt` in non-TLS-enabled builds
now that glibc crypt support is optional[0].

Additionally, many of the `nativeBuildInputs` should be `buildInputs`
since they are used at runtime and not build time. This is also fixed in
this commit.

[0]: ff30c899d8bd92d1a1c9f4d4e81455b04cb0868e