sane_backends: more complete hwdb files
We used to ship the pre-computed udev rules that are distributed in the
tarballs. This is problematic as it changes the group of scanners to
scanner which removes the group lp and prevents cups from using it.
(https://github.com/NixOS/nixpkgs/issues/147217)
For this reason we switched to generating udev and hwdb files as follows:

    ./tools/sane-desc -m udev+hwdb -s doc/descriptions:doc/descriptions-external > $out/etc/udev/rules.d/49-libsane.rules
    ./tools/sane-desc -m udev+hwdb -s doc/descriptions -m hwdb > $out/etc/udev/hwdb.d/20-sane.hwdb

following what other distros do.
(https://github.com/NixOS/nixpkgs/pull/176412)

this is great, but sane-desc does not include unsupported .desc entries
like EPSON V300 PHOTO, which was present in the distributed udev rules.
So even when installing all the required unfree stuff to make the
corresponding epkowa backend work, it only works as root as there is no
corresponding hwdb entry. As the .desc entry contains the usb ids, we
can just patch sane-desc to generate hwdb entries even for unsupported
models.