signal-desktop-beta: reinit at 6.1.0-beta.1
signal-desktop-beta was removed due to maintainability concerns in
256e347dfe2e8ab0070d85320fc1cc7da163eb92. Unfortunately, there are a
number of features, like stories, that are only enabled on the beta
builds. As such, this change adds that package back, but based upon the
current signal-desktop derivation.

We have restructured the derivation into the generic+versions structure
seen in libmicrohttpd, however since there are a fixed number of
releases, stable and beta, they both live default.nix and are inherited
at the top level.

I have also added myself to the maintainers list, however we could also
split that up, so that I alone bear the maintenance responsibilities.