tree-sitter: remove `language` name changes
The `language` attr is only used for the grammar name, changing it
makes no sense at all, and makes the output names irregular.

I read something about nvim inferring the language from the filepath
or something, I seriously hope that’s not A Thing, and if it is, we
should find a different way of solving that.

The tree-sitter integration is not specific to nvim, so it shouldn’t
have nvim hacks.