mudlet: 4.15.1 -> 4.16.0
This commit updates Mudlet from version 4.15.1 to 4.16.0. This requires
fixing a lua 5.1 compatibility regression in luasql-sqlite3 by using
a packageOverride to set the version to a commit before the regression.
This resolves an "undefined symbol: lua_isinteger" error that will
occur at application runtime otherwise.

Similarly, prior to this commit running Mudlet would produce errors in
the console window related to not finding the required `.so` for the
Discord integration to function.

This commit also wires through the required `libdiscord-rpc.so` from the
discord-rpc package. On startup Mudlet now prints:
> Discord integration loaded. Using functions from: "libdiscord-rpc.so"