boehmgc: refine test-disablement
Commit 172f9cf80723482ab506365902a21025963bf4da disabled the checks
for boehmgc on darwin-x86_64 with the comment "`gctest` fails under
emulation on aarch64-darwin".

Presumably this is because the test does not work under Apple's
Rosetta binary translator, which is used to execute x86_64 binaries
on aarch64 hosts.  The `check` phase would only use this translator
if a narrower condition is met:

```
stdenv.isDarwin && stdenv.buildPlatform.isAarch64 && stdenv.hostPlatform.isx86_64
```

Let's use that condition instead, mainly as documentation so people
don't think that the boehm-gc test suite is unreliable (this
happened already in https://github.com/NixOS/nixpkgs/pull/198591).
The boehm-gc test suite is actually quite good, and it is extremely
important to run the tests for this particular package because of
the number of advanced conservative-collector techniques it uses.
Malfunctions here can cause use-after-`free()` bugs that don't
correspond to any particular invocation of `free()` the way they
would in a C/C++ program.  They are incredibly difficult to track
down.