Update vmware-workstation to 16.2.4 and support MacOS Guests
The purpose of this change is to both update VMWare Workstation to v16.2.4 (latest version at this time) and allow macOS guests to be run on top of non-Apple host systems.

This functionality is provided on other distros such as ArchLinux & Gentoo by AUR packages or USE flags in gentoo's case. This will be a useful addition and I am happy to maintain it (and as such have added myself as a maintainer of the package.)