mame-tools: init at 0.249
This commit adds a new `tools` output to `mame` and move all
non-emulator binaries to it. This is because some of those binaries
(e.g.: chdman) are useful outside of the MAME, and the MAME binary is
really big. So if you only want to use some of the MAME tools, this will
reduce the download size needed for it.

The `tools` output is declared as `mame-tools` inside
`all-packages.nix`, for easier discovery.

This is similar to how other distros packages MAME, e.g.: on Arch Linux
we have `mame` and `mame-tools`.