pfstools: fix netpbm inclusion
The main CMakeLists.txt no longer includes the FindNETPBM.cmake file.
This commit removes the comment on the `include` statement to re-enable
including the FindNETPBM.cmake file.

Moreover the netpbm libraries weren't found due to the name searched for
included an extraneous `.` The
`stdenv.hostPlatform.extensions.sharedLibrary` expressions resolves to
".so" on x86_64-linux which means that the `-name` flag gets passed
"*..so*" instead of "*.so*". This commit ensures that "*.so*" is passed
when looking for the netpbm shared objects.