pkgsStatic.protobufc: fix build
Upstream effectively puts "-lprotobuf -lprotoc" as linking flags, while in fact
"protoc.a" depends on "protobuf.a", so flags should be in reverse order.

It is simpler to append one more "-lprotobuf" than to make patch to fix order.
That means that linker will scan "protobuf.a" twice, but price is neglectable.