emacs.pkgs.ada-mode: fix build for version 7.3.1
Fixes breakage to ada-mode introduced in #200190.

ada-mode recently gained a Makefile [1] which is intended for compiling
it in the ELPA worktree. Building from the ELPA release tarball still
needs to be done via build.sh, so we need to overwrite buildPhase to
ignore the Makefile (previously it'd just do nothing).

Additionally some new build tools are needed and lzip can be removed (as
we no longer fetch an archived version of wisi from ELPA).

[1]: https://git.savannah.gnu.org/cgit/emacs/elpa.git/commit/Makefile?h=externals/ada-mode&id=d5428433e39f3de7168fdc891127a2dcff849904