php80.packages.php-parallel-lint: Fix build
One of the main use cases of this program is ensuring that a project
targeting older PHP version does not use syntax that is not available there.

Unfortunately, box is now only available for PHP ≥ 8.1.
And even if we bring it from PHP 8.1 package set, the produced PHAR
will contain syntax incompatible with older PHP versions.

To work around this we either need to disable requirements checks
or force use the default PHAR stub over the box’s one.
And it turns out that php-parallel-linter already does the former,
only we did not have access to the `box.json` file
because it was excluded from the `git archive`.

Let’s fetch the source using Git so that box configuration is available.

Reverts https://github.com/NixOS/nixpkgs/commit/2db0ddb968357051237915256b7a1fc841492762