pkgs/stdenv: normalize attributes where setup.sh makes no distinction
This PR normalizes the final attrset immediately before it is passed
to `builtins.derivation`, deleting attributes whose value is `[]`,
`""`, or `false` if `setup.sh` treats that particular value for that
particular attribute indentically to the attribute being absent.

As a result of this PR, it is possible to add, to an existing
derivation, an attribute with the form:

```
  hardeningFlags = lib.optionals stdenv.hostPlatform.isFrobnicator [
    ...
```

without triggering an unnecessary mass-rebuild unless Frobnicator is
one of the Hydra `supportedPlatforms`.  Previously it was necessary
to either use the non-idiomatic:

```
  if !stdenv.hostPlatform.isFrobnicator then null else [
```

or to trigger a mass-rebuild and suffer OfBorg's "red ovals of doom"
on the PR.  The latter case tends to lead to stale and forgotten PRs.

Recent example: https://github.com/NixOS/nixpkgs/pull/199978#discussion_r1021076072