CONTRIBUTING: Reference release notes in package bumps
This change improves the recommendation for good commit messages to
include release notes on package bumps.

Including the release notes increases the likelihood that they will be
taken into consideration during review. Having them included in the
review is important to be able to judge whether changes made during a
version bump are sensible, sufficient or complete.

The burden of retrieving the release notes for arbitrary package bumps
should not rest on the relatively small group of reviewers.

Signed-off-by: Martin Weinelt <hexa@darmstadt.ccc.de>